fix-apps-slug-pages (#104)
* fix-apps-slug-pages

* fix

* fix

* fix

* fix
