+++ packages/app-store/routing-forms/pages/route-builder/[...appPages].tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { useAutoAnimate } from "@formkit/auto-animate/react";
 import Link from "next/link";
 import React, { useCallback, useState } from "react";
