+++ packages/app-store/routing-forms/pages/routing-link/[...appPages].tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import Head from "next/head";
 import { useRouter, useSearchParams } from "next/navigation";
 import type { FormEvent } from "react";
@@ -6,22 +8,20 @@ import { Toaster } from "react-hot-toast";
 import { v4 as uuidv4 } from "uuid";
 
 import { sdkActionManager, useIsEmbed } from "@calcom/embed-core/embed-iframe";
-import { orgDomainConfig } from "@calcom/features/ee/organizations/lib/orgDomains";
 import classNames from "@calcom/lib/classNames";
 import useGetBrandingColours from "@calcom/lib/getBrandColours";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
 import useTheme from "@calcom/lib/hooks/useTheme";
 import { trpc } from "@calcom/trpc/react";
-import type { AppGetServerSidePropsContext, AppPrisma } from "@calcom/types/AppGetServerSideProps";
 import type { inferSSRProps } from "@calcom/types/inferSSRProps";
 import { Button, showToast, useCalcomTheme } from "@calcom/ui";
 
 import FormInputFields from "../../components/FormInputFields";
 import getFieldIdentifier from "../../lib/getFieldIdentifier";
-import { getSerializableForm } from "../../lib/getSerializableForm";
 import { processRoute } from "../../lib/processRoute";
 import transformResponse from "../../lib/transformResponse";
 import type { Response, Route } from "../../types/types";
+import { getServerSideProps } from "./getServerSideProps";
 
 type Props = inferSSRProps<typeof getServerSideProps>;
 const useBrandColors = ({
@@ -232,68 +232,7 @@ export default function RoutingLink(props: inferSSRProps<typeof getServerSidePro
 
 RoutingLink.isBookingPage = true;
 
-export const getServerSideProps = async function getServerSideProps(
-  context: AppGetServerSidePropsContext,
-  prisma: AppPrisma
-) {
-  const { params } = context;
-  if (!params) {
-    return {
-      notFound: true,
-    };
-  }
-  const formId = params.appPages[0];
-  if (!formId || params.appPages.length > 2) {
-    return {
-      notFound: true,
-    };
-  }
-  const { currentOrgDomain, isValidOrgDomain } = orgDomainConfig(context.req);
-
-  const isEmbed = params.appPages[1] === "embed";
-
-  const form = await prisma.app_RoutingForms_Form.findFirst({
-    where: {
-      id: formId,
-      user: {
-        organization: isValidOrgDomain
-          ? {
-              slug: currentOrgDomain,
-            }
-          : null,
-      },
-    },
-    include: {
-      user: {
-        select: {
-          username: true,
-          theme: true,
-          brandColor: true,
-          darkBrandColor: true,
-        },
-      },
-    },
-  });
-
-  if (!form || form.disabled) {
-    return {
-      notFound: true,
-    };
-  }
-
-  return {
-    props: {
-      isEmbed,
-      themeBasis: form.user.username,
-      profile: {
-        theme: form.user.theme,
-        brandColor: form.user.brandColor,
-        darkBrandColor: form.user.darkBrandColor,
-      },
-      form: await getSerializableForm({ form }),
-    },
-  };
-};
+export { getServerSideProps };
 
 const usePrefilledResponse = (form: Props["form"]) => {
   const searchParams = useSearchParams();
